win32 fixes
git-svn-id: https://workrave.svn.sourceforge.net/svnroot/workrave/workrave/workrave/trunk@929 2b0e655b-6f86-497a-824e-95ef0e1e0d90

+++ common/src/win32/crashlog.c
@@ -464,7 +464,7 @@ save_key(FILE *log, HKEY key, char *name)
         }
       else if (type == REG_DWORD && data_size==4)
         {
-          fprintf(log, "  dword data = %08lx\n", data);
+          fprintf(log, "  dword data = %08lx\n", (long)data);
         }
       else
         {
